Convert EMotionFX runtime uint32 -> size_t
This allows the EMotionFX runtime to compile with `/we4267` enabled, which
emits a warning when converting from `size_t` to a smaller type. All tests
for the runtime have been updated accordingly, and they pass.

In instances where a range-for loop could be used, or a std algorithm, that
was used instead of using `size_t numItems = vec.size()` and a for loop.

Casts to `uint32` were removed where possible. Some places remain, like in
the file formats.

+++ Gems/EMotionFX/Code/EMotionFX/CommandSystem/Source/ActorCommands.h
@@ -19,9 +19,9 @@ namespace CommandSystem
 {
     // Adjust the given actor.
     MCORE_DEFINECOMMAND_START(CommandAdjustActor, "Adjust actor", true)
-    uint32                                          mOldMotionExtractionNodeIndex;
-    uint32                                          mOldRetargetRootNodeIndex;
-    uint32                                          mOldTrajectoryNodeIndex;
+    size_t                                          mOldMotionExtractionNodeIndex;
+    size_t                                          mOldRetargetRootNodeIndex;
+    size_t                                          mOldTrajectoryNodeIndex;
     AZStd::string                                   mOldAttachmentNodes;
     AZStd::string                                   mOldExcludedFromBoundsNodes;
     AZStd::string                                   mOldName;
@@ -71,6 +71,6 @@ public:
     // Helper functions
     //////////////////////////////////////////////////////////////////////////////////////////////////////////
     void COMMANDSYSTEM_API ClearScene(bool deleteActors = true, bool deleteActorInstances = true, MCore::CommandGroup* commandGroup = nullptr);
-    void COMMANDSYSTEM_API PrepareCollisionMeshesNodesString(EMotionFX::Actor* actor, uint32 lod, AZStd::string* outNodeNames);
+    void COMMANDSYSTEM_API PrepareCollisionMeshesNodesString(EMotionFX::Actor* actor, size_t lod, AZStd::string* outNodeNames);
     void COMMANDSYSTEM_API PrepareExcludedNodesString(EMotionFX::Actor* actor, AZStd::string* outNodeNames);
 } // namespace CommandSystem
